requirements.txt
jsonschema>=4.23,<5
python-docx>=1.1,<2
